Reorganize file structure for repo
