+++ package.json
@@ -12,7 +12,7 @@
     "test": "jest",
     "build": "webpack --config webpack.config.js --watch",
     "start": "nodemon server/index.js",
-    "schema": "mysql -u root < server/schema.sql"
+    "schema": "mysql -u root < server/database/mysql/schema.sql"
   },
   "dependencies": {
     "body-parser": "^1.19.0",
