+++ server/database/mysql/schema.sql
